refactor: change the directory into subfolders fix(Features): add 'api' guards to the auth argument to fix auth()->user() intelephense error
