+++ backend/app/Http/Controllers/Features/ReportController.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Features;
 
+use App\Http\Controllers\Controller;
+use App\Http\Requests\Features\ReportRequest;
 use App\Models\Report;
-use App\Http\Requests\ReportRequest;
 use App\Http\Resources\ReportResource;
 
 class ReportController extends Controller
@@ -24,8 +25,7 @@ class ReportController extends Controller
      */
     public function store(ReportRequest $request)
     {
-        // Kalau merah biarin aja, masih tetep jalan. Extension Intelephense ga bisa ngedeteksi method user()
-        $report = auth()->user()->reports()->create($request->validated());
+        $report = auth('api')->user()->reports()->create($request->validated());
 
         return response()->json([
             'message' => 'Create report success.',
